+++ projects/ngx-feature-toggle/src/lib/ngx-feature-toggle-route-guard.router.ts
@@ -1,31 +1,34 @@
 import { Injectable, isDevMode } from '@angular/core';
 import {
   ActivatedRouteSnapshot,
-  CanActivate,
   Route,
+  CanActivateChild,
   Router,
+  CanLoad,
+  CanActivate,
 } from '@angular/router';
 import { isOn } from 'feature-toggle-service';
 
 @Injectable({ providedIn: 'root' })
-export class NgxFeatureToggleCanActivateGuard implements CanActivate {
+export class NgxFeatureToggleRouteGuard
+  implements CanActivateChild, CanLoad, CanActivate {
   constructor(private router: Router) {}
 
   isDevMode() {
     return isDevMode();
   }
 
-  canActivate(route: ActivatedRouteSnapshot | Route): boolean {
-    const DEV_MODE = this.isDevMode();
+  private isOnCheck(route: ActivatedRouteSnapshot | Route): boolean {
     if (
       !route ||
       !route.data ||
       (typeof route.data.featureToggle !== 'string' &&
         !Array.isArray(route.data.featureToggle))
     ) {
-      if (DEV_MODE) {
+      if (this.isDevMode()) {
         console.error(
-          '`NgxFeatureToggleCanActivateGuard` need to receive `featureToggle` as data as an array or string in your route configuration.'
+          // tslint:disable-next-line: max-line-length
+          '`NgxFeatureToggleRouteGuard` need to receive `featureToggle` as data as an array or string in your route configuration.'
         );
       }
       return false;
@@ -42,5 +45,17 @@ export class NgxFeatureToggleCanActivateGuard implements CanActivate {
     }
 
     return hasAllTogglesOn;
+  }
+
+  canLoad(route: Route): boolean {
+    return this.isOnCheck(route);
+  }
+
+  canActivateChild(route: ActivatedRouteSnapshot | Route): boolean {
+    return this.isOnCheck(route);
+  }
+
+  canActivate(route: ActivatedRouteSnapshot | Route): boolean {
+    return this.isOnCheck(route);
   }
 }