+++ projects/ngx-feature-toggle/src/public-api.ts
@@ -5,6 +5,4 @@
 export * from './lib/ngx-feature-toggle.directive';
 export * from './lib/ngx-feature-toggle-provider.component';
 export * from './lib/ngx-feature-toggle.module';
-export * from './lib/ngx-feature-toggle-can-activate-guard.router';
-export * from './lib/ngx-feature-toggle-can-activate-child-guard.router';
-export * from './lib/ngx-feature-toggle-can-load-guard.router';
+export * from './lib/ngx-feature-toggle-route-guard.router';