+++ src/app/app-routing.module.ts
@@ -1,11 +1,9 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { NgxFeatureToggleCanActivateGuard } from '../../projects/ngx-feature-toggle/src/lib/ngx-feature-toggle-can-activate-guard.router';
+import { NgxFeatureToggleRouteGuard } from '../../projects/ngx-feature-toggle/src/lib/ngx-feature-toggle-route-guard.router';
 import { ErrorComponent } from './error/error.component';
 import { HomeComponent } from './home/home.component';
 import { set } from 'feature-toggle-service';
-// tslint:disable-next-line: max-line-length
-import { NgxFeatureToggleCanActivateChildGuard } from '../../projects/ngx-feature-toggle/src/lib/ngx-feature-toggle-can-activate-child-guard.router';
 import { CustomerComponent } from './customer/customer.component';
 import { CustomerDetailComponent } from './customer/customer-detail.component';
 import { RestrictPageDueFeatureToggleComponent } from './restrict-page-due-feature-toggle/restrict-page-due-feature-toggle.component';
@@ -21,7 +19,8 @@ export const routes: Routes = [
   {
     path: 'home',
     component: HomeComponent,
-    canActivate: [NgxFeatureToggleCanActivateGuard],
+    canLoad: [NgxFeatureToggleRouteGuard],
+    canActivate: [NgxFeatureToggleRouteGuard],
     data: {
       featureToggle: ['enableSecondText'],
       redirectTo: '/error',
@@ -30,7 +29,7 @@ export const routes: Routes = [
   {
     path: 'restrict',
     component: RestrictPageDueFeatureToggleComponent,
-    canActivate: [NgxFeatureToggleCanActivateGuard],
+    canActivate: [NgxFeatureToggleRouteGuard],
     data: {
       featureToggle: ['!enableSecondText'],
       redirectTo: '/error',
@@ -39,7 +38,7 @@ export const routes: Routes = [
   {
     path: 'error',
     component: ErrorComponent,
-    canActivate: [NgxFeatureToggleCanActivateGuard],
+    canActivate: [NgxFeatureToggleRouteGuard],
     data: {
       featureToggle: ['enableFirstText'],
     },
@@ -47,8 +46,8 @@ export const routes: Routes = [
   {
     path: 'customer',
     component: CustomerComponent,
-    canActivate: [NgxFeatureToggleCanActivateGuard],
-    canActivateChild: [NgxFeatureToggleCanActivateChildGuard],
+    canActivate: [NgxFeatureToggleRouteGuard],
+    canActivateChild: [NgxFeatureToggleRouteGuard],
     data: {
       featureToggle: ['enableCustomerPage'],
       redirectTo: '/error',